# docs/API_documentation.yaml
openapi: '3.0.2'
info:
  title: Kdo_Danes_Pospravlja
  version: '1.0'
servers:
  - url: https://localhost:4000/
paths: 
...
  /units:
    
    get:
      
      summary: "Pridobitev vseh enot"
      
      responses:
        
        '200':
          description: "Vse enote, ki so vezane za posameznega registriranega uporabnika"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: array
                    items:
                      type: object
                      properties:
                        unit_id:
                          type: integer
                        unit_name:
                          type: string
                          default: "Ime bivalne enote"
                        street:
                          type: string
                          default: "Naziv ceste"
                        place:
                          type: string
                          default: "Naziv pošte"
                        postal_code:
                          type: integer
                        country:
                          type: string
                          default: "Naziv države"
                        pictures:
                          type: array
                          items:
                            type: string
                            example: ["Pot do slike 1", "Pot do slike 2", "Pot do slike 3"]
                        roommates:
                          type: array
                          items:
                            type: object
                            properties:
                              name:
                                type: string
                                example: "Janez Oseba"
                              usernmae:
                                type: string
                                example: "janez.oseba"
                              avatar:
                                type: string
                                example: "./img/avatars/man-3.png"
                        spaces:
                          type: array
                          items:
                            type: string
                            example: ["kuhinja", "kopalnica", "klet"]
                              
        '500':
          description: "Napaka na strežniku"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    default: "Napaka na strežniku. Prosimo, poskusite kasneje."
  
    post:
    
      summary: "Ustvarjanje nove enote"
    
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                unit_name:
                  type: string
                  default: "Ime bivalne enote"
                street:
                  type: string
                  default: "Naziv ceste"
                place:
                  type: string
                  default: "Naziv pošte"
                postal_code:
                  type: integer
                country:
                  type: string
                  default: "Naziv države"
                rooms:
                  type: array
                  items:
                    type: string
                    example: ["Kuhinja", "Kopalnica", "Dvorišče"]
                    description: "Nazivi prostorov v bivalni enoti"
                pictures:
                  type: array
                  items:
                    type: string
                    example: ["Pot do slike 1", "Pot do slike 2", "Pot do slike 3"]
        
      
      responses:
        
        '200':
          description: "Uspešno dodajanje"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    default: "Bivalna enota je uspešno dodana."
        
        '400':
          description: "Napaka na odjemalcu"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    default: "Prosimo, še enkrat vpišite podatke."
        '500':
          description: "Napaka na strežniku"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    default: "Napaka na strežniku. Prosimo, poskusite kasneje."
  /units/{unit_id}:
    
    get:
    
      summary: "Pridobitev specifične bivalne enote glede na id"
    
      parameters:
        - name: unit_id
          in: path
          required: true
          schema:
            type: integer
            
      responses:
        
        '200':
          description: "Posamezna bivalna enota"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: object
                    properties:
                        unit_id:
                          type: integer
                        unit_name:
                          type: string
                          default: "Ime bivalne enote"
                        street:
                          type: string
                          default: "Naziv ceste"
                        place:
                          type: string
                          default: "Naziv pošte"
                        postal_code:
                          type: integer
                        country:
                          type: string
                          default: "Naziv države"
                        pictures:
                          type: array
                          items:
                            type: string
                            example: ["Pot do slike 1", "Pot do slike 2", "Pot do slike 3"]
                        roommates:
                          type: array
                          items:
                            type: object
                            properties:
                              name:
                                type: string
                                example: "Janez Oseba"
                              usernmae:
                                type: string
                                example: "janez.oseba"
                              avatar:
                                type: string
                                example: "./img/avatars/man-3.png"
                        spaces:
                          type: array
                          items:
                            type: string
                            example: ["kuhinja", "kopalnica", "klet"]
        
        '401':
          description: "JWT ni poslan"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    default: "Ne morete dostopati do te bivalne enote."
        
        '404':
          description: "Bivalna enota ni najdena"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    default: "Žal nam je, bivalna enota, ki jo iščete, ni bila najdena."
        
        '500':
          description: "Napaka na strežniku"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    default: "Napaka na strežniku. Prosimo, poskusite kasneje."
    
    put:
      
      summary: "Posodobitev določene enote"
      
      parameters:
        - name: unit_id
          in: path
          required: true
          schema:
            type: integer
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                unit_name:
                  type: string
                  default: "Ime bivalne enote"
                street:
                  type: string
                  default: "Naziv ceste"
                place:
                  type: string
                  default: "Naziv pošte"
                postal_code:
                  type: integer
                country:
                  type: string
                  default: "Naziv države"
                rooms:
                  type: array
                  items:
                    type: string
                    example: ["Kuhinja", "Kopalnica", "Dvorišče"]
                    
      responses:
        
        '200':
          description: "Uspešna posodobitev"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: object
                    default: "Bivalna enota je uspešno posodobljena."
        
        '400':
          description: "Napaka na odjemalcu"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    default: "Prosimo, še enkrat vpišite podatke."
        
        '404':
          description: "Bivalna enota ni najdena"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    default: "Žal nam je, bivalna enota, ki jo iščete, ni bila najdena."
        
        '500':
          description: "Napaka na strežniku"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    default: "Napaka na strežniku. Prosimo, poskusite kasneje."

    delete:
      
        summary: "Odstranitev bivalne enote"
      
        parameters:
          - name: unit_id
            in: path
            required: true
            schema:
              type: integer
          - name: username
            in: path
            required: true
            schema:
              type: string
        
        responses:
          
          '200':
            description: "Odstranitev uspešna"
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    msg:
                      type: string
                      default: "Bivalna enota je uspešno odstranjena"
                      
          '401':
            description: "JWT ni poslan."
                      
          '403':
            description: "Uporabnik nima dostopa do odstranjevanja enote"
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    msg:
                      type: string
                      default: "Nimate dostopa za odstranitev enote."
          
          '404':
            description: "Enota ni najdena"
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    msg:
                      type: string
                      default: "Žal nam je, ta enota ni najdena."
          
          '500':
            description: "Napaka na strežniku"
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    msg:
                      type: string
                      default: "Napaka na strežniku. Prosimo, poskusite kasneje."
  /units/{unit_id}/users:
    
    get:
      
      summary: "Pridobitev uporabnikov, prijavljenih na bivalno enoto"
      
      parameters:
        - name: unit_id
          in: path
          required: true
          schema:
            type: integer
      
      responses:
        
        '200':
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        username:
                          type: string
                        name:
                          type: string
                        surname:
                          type: string
                          
        '500':
          description: "Napaka na strežniku"
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    default: "Napaka na strežniku. Prosimo, poskusite kasneje."
...